Segments list of containers into new, old and duplicates
Queries docker to get a list of containers.  If the container is already in
etcd, the config is not changed.  If the container is new it is added.  If a
etcd still has a container registered but it doesn't seem to be running, the
container is put in a list to be deregistered (TODO).  Likewise, if there is
a duplicate registration in etcd, the container is registered to be deleted.